requirements.txt
discord.py>=2.3.0
aiosqlite
python-dotenv
